Organize render module
