+++ raster/src/main/scala/geotrellis/raster/render/JpgRenderMethods.scala
@@ -2,7 +2,6 @@ package geotrellis.raster.render
 
 import geotrellis.raster._
 import geotrellis.raster.render.jpg._
-import geotrellis.raster.render.shared._
 import geotrellis.raster.histogram.Histogram
 import geotrellis.raster.op.stats._
