+++ raster/src/main/scala/geotrellis/raster/render/PngRenderMethods.scala
@@ -2,7 +2,6 @@ package geotrellis.raster.render
 
 import geotrellis.raster._
 import geotrellis.raster.render.png._
-import geotrellis.raster.render.shared._
 import geotrellis.raster.histogram.Histogram
 import geotrellis.raster.op.stats._
