+++ raster/src/main/scala/geotrellis/raster/render/color/Color.scala
