+++ raster/src/main/scala/geotrellis/raster/render/color/ColorMap.scala
