+++ raster/src/main/scala/geotrellis/raster/render/color/ColorRamp.scala
