+++ raster/src/main/scala/geotrellis/raster/render/color/ColorRamps.scala
