+++ raster/src/main/scala/geotrellis/raster/render/package.scala
@@ -2,5 +2,5 @@ package geotrellis.raster
 
 package object render {
   implicit class RenderMethodExtensions(val tile: Tile) extends SharedRenderMethods with PngRenderMethods with JpgRenderMethods { }
-  implicit class MultiBandRenderMethodExtensions(val tile: MultiBandTile) extends MultiBandRenderMethods { }
+  implicit class MultiBandRenderMethodExtensions(val tile: MultiBandTile) extends MultiBandPngRenderMethods { }
 }
